+++ Kernel/System/Ticket/Article/Backend/Email.pm
@@ -390,6 +390,7 @@ sub ArticleBounce {
         From         => $Param{From},
         To           => $Param{To},
         Email        => $Email,
+        ArticleID    => $Param{ArticleID},
     );
 
     return if !$BounceSent->{Success};